Create Adblock Plus Whitelist
Adblock Plus, while a great tool, sometimes needs things whitelisted across a domain. Since Adblock Plus is silly and doesn't provide that functionality via a GPO extension, here's the workaround. On a test machine, whitelist everything you need to in Adblock Plus and save it. Then, use $PathLocation line in the code to find out where the whitelist file is located at. Copy it to a central location and make $ServerPath point to it. Voila. Now you can constantly push out a whitelist of affected sites to your Chrome users.